add submodule model_managers, adapt imports
