+++ gnn_aid/attacks/metattack/meta_gradient_attack.py
@@ -13,7 +13,7 @@ from gnn_aid.attacks.poison_attacks import PoisonAttacker
 from gnn_aid.aux.utils import OPTIMIZERS_PARAMETERS_PATH, move_to_same_device
 from gnn_aid.data_structures.configs import ModelModificationConfig, ConfigPattern
 from gnn_aid.datasets.gen_dataset import GeneralDataset
-from gnn_aid.models_builder.gnn_models import FrameworkGNNModelManager, GNNModelManager
+from gnn_aid.models_builder.model_managers import GNNModelManager, FrameworkGNNModelManager
 from gnn_aid.models_builder.models_zoo import model_configs_zoo
 from . import utils
