+++ gnn_aid/models_builder/__init__.py
@@ -1,30 +1,28 @@
-from .attack_defense_manager import FrameworkAttackDefenseManager
-from .custom_layers import GSATLayer, ProtLayer, ExtractorMLP, GSATMLP
 from .gnn_constructor import FrameworkGNNConstructor, GNNConstructor, GNNConstructorTorch
-from .gnn_models import (
-    GNNModelManager, FrameworkGNNModelManager, ProtGNNModelManager, GSATModelManager)
 from .models_utils import (
     apply_attention_to_messages, attention_message_hook, apply_message_gradient_capture,
-    apply_decorator_to_graph_layers, EdgeMaskingWrapper)
+    apply_decorator_to_graph_layers, EdgeMaskingWrapper, Metric)
 from .models_zoo import model_configs_zoo
+from .custom_layers import GSATLayer, ProtLayer, ExtractorMLP, GSATMLP
+
+import gnn_aid.models_builder.model_managers
+
+from .attack_defense_manager import FrameworkAttackDefenseManager
 
 __all__ = [
-    "FrameworkAttackDefenseManager",
-    "GSATLayer",
-    "ProtLayer",
-    "ExtractorMLP",
-    "GSATMLP",
-    "FrameworkGNNConstructor",
-    "GNNConstructor",
-    "GNNConstructorTorch",
-    "GNNModelManager",
-    "FrameworkGNNModelManager",
-    "ProtGNNModelManager",
-    "GSATModelManager",
-    "apply_attention_to_messages",
-    "attention_message_hook",
-    "apply_message_gradient_capture",
-    "apply_decorator_to_graph_layers",
-    "EdgeMaskingWrapper",
-    "model_configs_zoo",
+    'FrameworkAttackDefenseManager',
+    'GSATLayer',
+    'ProtLayer',
+    'ExtractorMLP',
+    'GSATMLP',
+    'FrameworkGNNConstructor',
+    'GNNConstructor',
+    'GNNConstructorTorch',
+    'apply_attention_to_messages',
+    'attention_message_hook',
+    'apply_message_gradient_capture',
+    'apply_decorator_to_graph_layers',
+    'EdgeMaskingWrapper',
+    'Metric',
+    'model_configs_zoo',
 ]
