+++ gnn_aid/models_builder/model_managers/__init__.py
@@ -1,0 +1,11 @@
+from .model_manager import GNNModelManager
+from .framework_mm import FrameworkGNNModelManager
+from .gsat_mm import GSATModelManager
+from .protgnn_mm import ProtGNNModelManager
+
+__all__ = [
+    'GNNModelManager',
+    'FrameworkGNNModelManager',
+    'ProtGNNModelManager',
+    'GSATModelManager',
+]
