+++ tests/explainers_full_test.py
@@ -13,7 +13,8 @@ from gnn_aid.data_structures.configs import DatasetConfig, DatasetVarConfig, Con
 from gnn_aid.datasets.datasets_manager import DatasetManager
 from gnn_aid.datasets.ptg_datasets import LibPTGDataset
 from gnn_aid.explainers.explainers_manager import FrameworkExplainersManager
-from gnn_aid.models_builder.gnn_models import FrameworkGNNModelManager, ProtGNNModelManager, Metric
+from gnn_aid.models_builder.models_utils import Metric
+from gnn_aid.models_builder.model_managers import FrameworkGNNModelManager, ProtGNNModelManager
 from gnn_aid.models_builder.models_zoo import model_configs_zoo
 from tests.utils import cleanup_patches, monkey_patch_dirs
