+++ tests/explainers_test.py
@@ -9,7 +9,8 @@ from gnn_aid.datasets.datasets_manager import DatasetManager
 from gnn_aid.datasets.ptg_datasets import LibPTGDataset
 from gnn_aid.explainers.explainers_manager import FrameworkExplainersManager
 from gnn_aid.data_structures.configs import FeatureConfig, Task
-from gnn_aid.models_builder.gnn_models import FrameworkGNNModelManager, ProtGNNModelManager, Metric, GSATModelManager
+from gnn_aid.models_builder.models_utils import Metric
+from gnn_aid.models_builder.model_managers import FrameworkGNNModelManager, ProtGNNModelManager, GSATModelManager
 from gnn_aid.data_structures.configs import DatasetConfig, DatasetVarConfig, ConfigPattern, ModelModificationConfig
 from gnn_aid.models_builder.models_zoo import model_configs_zoo
 from tests.utils import cleanup_patches, monkey_patch_dirs
