+++ tests/models_test.py
@@ -4,8 +4,9 @@ import unittest
 
 from gnn_aid.datasets.datasets_manager import DatasetManager
 from gnn_aid.datasets.ptg_datasets import LibPTGDataset
-from gnn_aid.models_builder.gnn_models import FrameworkGNNModelManager, ProtGNNModelManager, Metric, \
-    GSATModelManager
+from gnn_aid.models_builder.models_utils import Metric
+from gnn_aid.models_builder.model_managers import (
+    FrameworkGNNModelManager, ProtGNNModelManager, GSATModelManager)
 from gnn_aid.data_structures.configs import ModelModificationConfig, DatasetConfig, DatasetVarConfig, \
     ConfigPattern, FeatureConfig, Task
 from gnn_aid.models_builder.models_zoo import model_configs_zoo
